Backend/requirements.txt
requests
beautifulsoup4
gunicorn
flask
flask-cors
nltk
python-dotenv
